+++ index.js
@@ -15,20 +15,24 @@ const bot = mineflayer.createBot({
 });
 
 module.exports = bot;
-// Handlers
+
 // Function folder
 const files = fs.readdirSync("./src/functions").filter(file => file.endsWith(".js"));
-
 for (const file of files) {
 	const eventName = file.split(".")[0];
-	console.log(chalk.cyanBright(`${eventName}.js`) + chalk.greenBright(` has been loaded.`));
+	// eslint-disable-next-line no-unused-vars
+	const event = require(`./src/functions/${file}`);
+	console.log(chalk.cyanBright(`Function ${eventName}.js`) + chalk.greenBright(` has been loaded.`));
 }
 
-// Events
+// Events folder
 const eFiles = fs.readdirSync("./src/events").filter(file => file.endsWith(".js"));
+
 for (const file of eFiles) {
 	const eventName = file.split(".")[0];
-	console.log(chalk.magentaBright(`${eventName}.js`) + chalk.blue(` event handler has been loaded.`));
+	// eslint-disable-next-line no-unused-vars
+	const event = require(`./src/events/${file}`);
+	console.log(chalk.magentaBright(`Event ${eventName}.js`) + chalk.blue(` has been loaded.`));
 }
 
 bot.loadPlugin(pathfinder);