+++ src/events/ready.js
@@ -1,7 +1,6 @@
 const bot = require('../../index');
-const { username } = require('./config.json').username;
 const chalk = require('chalk');
 
 bot.on('ready', () => {
-	console.log(chalk.greenBright(`${username} is ready to connect to your server!`));
+	console.log(chalk.greenBright(`${process.env.bot_username} is ready to connect to your server!`));
 });